// src/server/modules/ai/assistant/ai.assistant.enums.ts
export enum AiAssistant {
	DEFAULT = 'DEFAULT',
	INDEX_CREW = 'INDEX_CREW',
	INDEX_CREW_INVITE = 'INDEX_CREW_INVITE',
	INDEX_CREW_USER = 'INDEX_CREW_USER',
	MINT_DASHBOARD = 'MINT_DASHBOARD',
	MINT_INFO = 'MINT_INFO',
	MINT_CONFIG = 'MINT_CONFIG',
	MINT_KEYSETS = 'MINT_KEYSETS',
	MINT_KEYSET_ROTATION = 'MINT_KEYSET_ROTATION',
	MINT_DATABASE = 'MINT_DATABASE',
	MINT_BACKUP = 'MINT_BACKUP',
	EVENT_LOG = 'EVENT_LOG',
}